+++ NEWS.md
@@ -15,6 +15,10 @@
 * `format_table()` now also formats columns named `SGPV` (second generation
   p-values) as p-values.
 
+## Bug fixes
+
+* Fixed issue in `get_data()` for results from `kruskal.test()`.
+
 # insight 0.19.0
 
 ## New supported models